+++ config/settings/base.py
@@ -440,9 +440,10 @@ ENABLED_BACKEND_PLUGINS = env.list('ENABLED_BACKEND_PLUGINS', None, [
     'example_backend_app',
 ])
 
-# General API settings
+# SODAR API settings
 SODAR_API_DEFAULT_VERSION = '0.1'
-SODAR_API_MEDIA_TYPE = 'application/vnd.bihealth.sodar+json'
+SODAR_API_ALLOWED_VERSIONS = [SODAR_API_DEFAULT_VERSION]
+SODAR_API_MEDIA_TYPE = 'application/vnd.bihealth.sodar-core+json'
 SODAR_API_DEFAULT_HOST = env.url(
     'SODAR_API_DEFAULT_HOST', 'http://0.0.0.0:8000')